# lib/api-zod/src/generated/types/uidEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UidEntry {
  uid: string;
  expiry_date: string;
  bluestack: boolean;
  adder_name: string;
}