# lib/api-zod/src/generated/types/uidListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UidEntry } from "./uidEntry";

export interface UidListResponse {
  success: boolean;
  data: UidEntry[];
}